requirements.txt
flask==3.0.0
flask-cors==4.0.0
face-recognition==1.3.0  # newest pip wheel available for Python 3.14
opencv-python==4.8.1.78
numpy>=1.26.0
Pillow>=10.1.0
requests>=2.31.0
python-dotenv==1.0.0
cryptography>=41.0.0
qrcode>=7.4.0
pycryptodome>=3.19.0
supabase>=0.6.0
user-agent>=0.1.10
geopy>=2.3.0
